Fix warnings about signedness

+++ tests/auto/addresscache/tst_addresscache.cpp
@@ -76,8 +76,8 @@ private slots:
         for (auto addr : {0x100, 0x100 + 9}) {
             const auto cached = cache.findSymbol(libfoo_a, addr);
             QVERIFY(cached.isValid());
-            QCOMPARE(cached.offset, 0x100);
-            QCOMPARE(cached.size, 10);
+            QCOMPARE(cached.offset, quint64(0x100));
+            QCOMPARE(cached.size, quint64(10));
             QCOMPARE(cached.symname, "Foo");
         }
         QVERIFY(!cache.findSymbol(libfoo_a, 0x100 + 10).isValid());
